% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{r}
\alias{r}
\title{Return Path to \code{r} Binary}
\usage{
r(usecat = FALSE)
}
\arguments{
\item{usecat}{Optional toggle to request output to stdout (useful in Makefiles)}
}
\value{
The path is returned as character variable. If the \code{usecat} option is
set the character variable is displayed via \code{\link{cat}} instead.
}
\description{
Return the path of the install \code{r} binary.
}
\details{
The test for Windows is of course superfluous as we have no binary for Windows.
Maybe one day...
}
\author{
Dirk Eddelbuettel
}

